+++ patzilla/access/__init__.py
@@ -1,13 +1,18 @@
 # -*- coding: utf-8 -*-
-# (c) 2017-2018 Andreas Motl <andreas.motl@ip-tools.org>
-from ConfigParser import NoOptionError
+# (c) 2017-2022 Andreas Motl <andreas.motl@ip-tools.org>
+import logging
+
+logger = logging.getLogger(__name__)
+
 
 def includeme(config):
 
+    datasources = []
     try:
         datasources = config.registry.application_settings.ip_navigator.datasources
+        logger.info('Enabling data sources: {}'.format(datasources))
     except:
-        raise NoOptionError('datasources', 'ip_navigator')
+        logger.warning('No data source configured')
 
     if 'ops' in datasources:
         config.include("patzilla.access.epo.ops.client")
@@ -28,4 +33,3 @@ def includeme(config):
     """
 
     config.include('.office')
-    config.scan()